feat: address generate-folders review feedback
