+++ examples/generate-folders-folder-parent/.expected/diff.patch
@@ -1,0 +1,30 @@
+diff --git a/resources.yaml b/resources.yaml
+index 521e29f..f426e02 100644
+--- a/resources.yaml
++++ b/resources.yaml
+@@ -9,3 +9,25 @@ spec:
+   config:
+   - Dev:
+     - Team1
++---
++apiVersion: resourcemanager.cnrm.cloud.google.com/v1beta1
++kind: Folder
++metadata:
++  name: dev
++  annotations:
++    cnrm.cloud.google.com/blueprint: 'krm-fn'
++spec:
++  folderRef:
++    external: parent-folder-id
++  displayName: Dev
++---
++apiVersion: resourcemanager.cnrm.cloud.google.com/v1beta1
++kind: Folder
++metadata:
++  name: dev.team1
++  annotations:
++    cnrm.cloud.google.com/blueprint: 'krm-fn'
++spec:
++  folderRef:
++    name: dev
++  displayName: Team1
