+++ examples/generate-folders-folder-parent/.krmignore
@@ -1,0 +1,1 @@
+.expected
