+++ examples/generate-folders-folder-parent/resources.yaml
@@ -1,0 +1,11 @@
+apiVersion: blueprints.cloud.google.com/v1alpha3
+kind: ResourceHierarchy
+metadata:
+  name: test-folder-parent
+spec:
+  parentRef:
+    kind: Folder
+    external: parent-folder-id
+  config:
+  - Dev:
+    - Team1
