+++ examples/generate-folders-namespaced/.krmignore
@@ -1,0 +1,1 @@
+.expected
