+++ examples/generate-folders-namespaced/resources.yaml
@@ -1,0 +1,12 @@
+apiVersion: blueprints.cloud.google.com/v1alpha3
+kind: ResourceHierarchy
+metadata:
+  name: test-ns
+  namespace: test-ns
+spec:
+  parentRef:
+    kind: Organization
+    external: test-organization
+  config:
+  - Dev:
+    - Team1
