+++ examples/generate-folders-simple/.expected/diff.patch
@@ -1,0 +1,63 @@
+diff --git a/resources.yaml b/resources.yaml
+index c6e0011..45602f1 100644
+--- a/resources.yaml
++++ b/resources.yaml
+@@ -12,3 +12,58 @@ spec:
+     - Team_2
+   - Prod:
+     - Team One
++---
++apiVersion: resourcemanager.cnrm.cloud.google.com/v1beta1
++kind: Folder
++metadata:
++  name: dev
++  annotations:
++    cnrm.cloud.google.com/blueprint: 'krm-fn'
++spec:
++  organizationRef:
++    external: test-organization
++  displayName: Dev
++---
++apiVersion: resourcemanager.cnrm.cloud.google.com/v1beta1
++kind: Folder
++metadata:
++  name: dev.team-one
++  annotations:
++    cnrm.cloud.google.com/blueprint: 'krm-fn'
++spec:
++  folderRef:
++    name: dev
++  displayName: Team One
++---
++apiVersion: resourcemanager.cnrm.cloud.google.com/v1beta1
++kind: Folder
++metadata:
++  name: dev.team-2
++  annotations:
++    cnrm.cloud.google.com/blueprint: 'krm-fn'
++spec:
++  folderRef:
++    name: dev
++  displayName: Team_2
++---
++apiVersion: resourcemanager.cnrm.cloud.google.com/v1beta1
++kind: Folder
++metadata:
++  name: prod
++  annotations:
++    cnrm.cloud.google.com/blueprint: 'krm-fn'
++spec:
++  organizationRef:
++    external: test-organization
++  displayName: Prod
++---
++apiVersion: resourcemanager.cnrm.cloud.google.com/v1beta1
++kind: Folder
++metadata:
++  name: prod.team-one
++  annotations:
++    cnrm.cloud.google.com/blueprint: 'krm-fn'
++spec:
++  folderRef:
++    name: prod
++  displayName: Team One
