+++ examples/generate-folders-simple/.krmignore
@@ -1,0 +1,1 @@
+.expected
