+++ examples/generate-folders-simple/resources.yaml
@@ -1,0 +1,14 @@
+apiVersion: blueprints.cloud.google.com/v1alpha3
+kind: ResourceHierarchy
+metadata:
+  name: test-simple
+spec:
+  parentRef:
+    kind: Organization
+    external: test-organization
+  config:
+  - Dev:
+    - Team One
+    - Team_2
+  - Prod:
+    - Team One
